requirements.txt
Django==4.0.3
django-filter==22.1
djangorestframework==3.13.1
pytest==6.2.4
pytest-django==4.5.2
json
requests==2.25.1